pyproject.toml
[project]
name = "genai-agent-orchestrator"
version = "0.1.0"
description = "Agent orchestration service for GenAI Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "app",
    "agents",
    "orchestration",
    "mcp_client",
    "schemas"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
